# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vernietigingscockpit API
  version: 0.1.0
  description: >
    API-contract voor de review-branch van de vernietigingscockpit. Het contract
    beschrijft de selectie-, beoordelings-, accorderings-, uitvoerings- en
    resultaatstappen zoals gemapt in de huidige frontend.
  contact:
    name: Beheerteam Vernietigingscockpit
    email: beheerteam-vernietigingscockpit@example.org
    url: https://example.org/vernietigingscockpit/issues
  license:
    name: EUPL-1.2
    url: https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
servers:
  - url: https://api.vernietigingscockpit.test/v1
    description: Ontwikkelomgeving
security:
  - bearerAuth: []
...
paths:
...
  /taken/{taakId}/taakuitvoeringen/{taakuitvoeringId}/selectie:
    post:
      operationId: startSelectie
      tags:
        - Taakuitvoeringen
      summary: Start selectie ophalen
      description: Maakt een snapshot van vernietigbare objecten uit de gekoppelde stekkers.
      parameters:
        - $ref: "#/components/parameters/TaakId"
        - $ref: "#/components/parameters/TaakuitvoeringId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartSelectieRequest"
      responses:
        "202":
          description: Selectie is gestart.
          headers:
            API-Version:
              $ref: "#/components/headers/ApiVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActieStatus"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /taken/{taakId}/taakuitvoeringen/{taakuitvoeringId}/vernietiging:
    post:
      operationId: startVernietiging
      tags:
        - Taakuitvoeringen
      summary: Start vernietiging uitvoeren
      parameters:
        - $ref: "#/components/parameters/TaakId"
        - $ref: "#/components/parameters/TaakuitvoeringId"
      responses:
        "202":
          description: Vernietiging is gestart.
          headers:
            API-Version:
              $ref: "#/components/headers/ApiVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActieStatus"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /taken/{taakId}/taakuitvoeringen/{taakuitvoeringId}/archivering:
    post:
      operationId: archiveerTaakuitvoering
      tags:
        - Taakuitvoeringen
      summary: Archiveer de afgeronde taakuitvoering
      parameters:
        - $ref: "#/components/parameters/TaakId"
        - $ref: "#/components/parameters/TaakuitvoeringId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveerTaakuitvoeringRequest"
      responses:
        "202":
          description: Archivering is gestart.
          headers:
            API-Version:
              $ref: "#/components/headers/ApiVersion"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActieStatus"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    ApiVersion:
      description: Volledige semantische versie van de API-implementatie.
      schema:
        type: string
        pattern: "^\\d+\\.\\d+\\.\\d+$"
        example: 0.1.0
  parameters:
    TaakId:
      name: taakId
      in: path
      description: Unieke identificatie van de taakdefinitie.
      required: true
      schema:
        type: string
        minLength: 1
    TaakuitvoeringId:
      name: taakuitvoeringId
      in: path
      description: Unieke identificatie van de taakuitvoering.
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
...
    NotFound:
      description: Resource niet gevonden.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: Foutmelding volgens RFC 7807.
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: https://api.example.org/problems/not-found
        title:
          type: string
          example: Resource niet gevonden
        status:
          type: integer
          minimum: 100
          maximum: 599
          example: 404
        detail:
          type: string
          example: De gevraagde taakuitvoering bestaat niet.
        instance:
          type: string
          format: uri-reference
...
    StartSelectieRequest:
      type: object
      properties:
        stekkerIds:
          type: array
          description: Optionele selectie van stekkers. Bij leeg of afwezig worden alle actieve stekkers gebruikt.
          items:
            type: string
    ActieStatus:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
          description: Identificatie van de gestarte actie of job.
          example: job-123
        status:
          type: string
          enum:
            - GEACCEPTEERD
            - BEZIG
        melding:
          type: string
          example: Actie is gestart.
...
    ArchiveerTaakuitvoeringRequest:
      type: object
      properties:
        toelichting:
          type: string
          example: Afgeronde vernietiging gearchiveerd.